te-backend/te/requirements.txt
alembic==1.10.2
anyio==3.6.2
autoflake==2.0.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
emails==0.6
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
orjson==3.8.9
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.2.0
premailer==3.10.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0
google-auth 
google-auth-oauthlib
google-auth-httplib2 
google-api-python-client